+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Translations: [简体中文](CHANGELOG.zh.md)
 
 * fix: Fixed a bug where CrossfadePainter, ResizePainter, ImageBitmapPainter and DrawablePainter
   cannot fill the canvas due to loss of floating point precision
+* new: AsyncImage Adds the keepContentNoneStartWhenDraw param
 * depend: Update to kotlin 2.1.10
 * depend: Update to compose multiplatform 1.8.0
 * depend: Update to android-gif-drawable 1.2.29